==> scripts/visualization/01_fingerprint_matrix.ipynb <==
import pandas as pd
from rdkit import Chem

with open("../../results/schembl_summs_v4_gpt_cleaned_eps_0.340_diff_20030_final.csv", 'r') as f:
    df = pd.read_csv(f)

df["fingerprint"] = df["smiles"].apply(lambda x: Chem.RDKFingerprint(Chem.MolFromSmiles(x)))
df.to_pickle("schembl_summs_v4_final_with_fingerprint.pkl")
