==> scripts/visualization/06_co_graph.ipynb <==
import pandas as pd
import numpy as np
from ast import literal_eval

df = pd.read_csv("05_co_counter.csv")

df = df.rename(columns={"Unnamed: 0": "node_pair", "0": "weight"})

df["node_pair"] = df["node_pair"].apply(literal_eval)
df["node_1"] = df["node_pair"].apply(lambda x: x[0])
df["node_2"] = df["node_pair"].apply(lambda x: x[1])



df_piv = df.pivot(index="node_2", columns="node_1", values="weight")
df_piv = df_piv.fillna(0)
df_piv = df_piv.astype(int)

arr = df_piv.to_numpy()
# convert to array
arr = np.tril(arr)

# place arr back in df_piv
df_piv = pd.DataFrame(arr, columns=df_piv.columns, index=df_piv.index)




edges = []
# iterate through lower triangle of df_piv and add edges to edges
for i in range(len(df_piv.columns)):
    for j in range(len(df_piv.index)):
        if df_piv.iloc[j, i] > 0:
            start = df_piv.columns[i]
            end = df_piv.index[j]
            weight = df_piv.iloc[j, i]
            if start.isnumeric():
                start = int(start)
            if end.isnumeric():
                end = int(end)
            edges.append((start, end, int(weight)))

# convert edges to csv
df_exp = pd.DataFrame(columns=["source", "target", "weight"])
df_exp["source"] = [edge[0] for edge in edges]
df_exp["target"] = [edge[1] for edge in edges]
df_exp["weight"] = [edge[2] for edge in edges]

df_exp.to_csv("graph_data.csv", index=False)
# df_exp.to_csv("graph_data.ncol", sep=" ", index=False, header=False)
