==> scripts/visualization/09_analyze_t_tests.ipynb <==
import pandas as pd

df_p = pd.read_csv('primary_label_t_test.csv')
df_c = pd.read_csv('co-occurrence_t_test.csv')

# count number of significant p-values
print('Primary Label T-Test')
print('Number of significant p-values: ', len(df_p[df_p['p_val'] < 0.05]))
print('Number of insignificant p-values: ', len(df_p[df_p['p_val'] >= 0.05]))

print('\nCo-occurrence T-Test')
print('Number of significant p-values: ', len(df_c[df_c['p_val'] < 0.05]))
print('Number of insignificant p-values: ', len(df_c[df_c['p_val'] >= 0.05]))

# print all insignificant p-values
for label in df_p[df_p['p_val'] >= 0.05]['label']:
    print(label)
